Demo PR (#1)
* initial refactoring and demo commit

* add video demo
