--- misc.py
@@ -1,4 +1,0 @@
-import psutil
-
-def free_ram():
-    return psutil.virtual_memory().available
