+++ requirements.txt
@@ -1,0 +1,9 @@
+datasets 
+torch 
+matplotlib 
+numpy
+streamlit
+transformers
+numpy
+scipy
+opencv-python
