+++ src/misc.py
@@ -1,0 +1,4 @@
+import psutil
+
+def free_ram():
+    return psutil.virtual_memory().available
